Use selectors from one file and reorder files
