--- src/components/charts/divider-line/divider-line.js
@@ -1,3 +1,0 @@
-import Component from './divider-line-component';
-
-export default Component;
